Convert to Xcode 12

+++ SFTPSplitView.m
@@ -7,7 +7,7 @@
 
 @implementation SFTPSplitView
 
-- ( float )dividerThickness
+- ( double )dividerThickness
 {
     return( 20.0 );
 }
